+++ src/script/reducer/record/list.js
@@ -1,0 +1,3 @@
+module.exports = (records) => {
+  return records.slice().reverse();
+};
